requirements.txt
dash
dash_bootstrap_components
plotly
pandas
python-docx
